router/services/tuning/requirements.txt
# Tuning Service Dependencies for Milestone 3 GBDT Pipeline
# ML Framework
lightgbm>=4.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data Processing
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Hyperparameter Optimization
optuna>=3.4.0
joblib>=1.3.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.25.0

# Storage and Serialization
boto3>=1.28.0  # for S3 artifact storage
cloudpickle>=2.2.0
tarfile-stream>=0.1.0

# Utilities
click>=8.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
tqdm>=4.66.0

# Monitoring
prometheus_client>=0.17.0
structlog>=23.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0